restructure & add compile/release stuff
