--- .github/workflows/check.yaml
@@ -1,14 +1,0 @@
-name: check
-on:
-    - push
-jobs:
-    check:
-        runs-on: ubuntu-latest
-        steps:
-            - uses: actions/checkout@v2
-            - uses: denoland/setup-deno@v1
-              with:
-                  deno-version: 1.15.2
-            - run: deno lint
-            - run: deno fmt --check
-            - run: deno test --unstable --allow-net --allow-run --lock=lock.json
