+++ .github/workflows/todo checker.yaml
@@ -1,18 +1,18 @@
 name: 'todo checker'
 on:
-    push:
-        branches: [master]
+  push:
+    branches: [ master ]
 jobs:
-    build:
-        runs-on: 'ubuntu-latest'
-        steps:
-            - uses: 'actions/checkout@master'
-            - name: 'TODO to Issue'
-              uses: 'alstr/todo-to-issue-action@master'
-              with:
-                  REPO: ${{ github.repository }}
-                  BEFORE: ${{ github.event.before }}
-                  SHA: ${{ github.sha }}
-                  TOKEN: ${{ secrets.GITHUB_TOKEN }}
-                  LABEL: 'TODO'
-              id: 'todo'
+  build:
+    runs-on: 'ubuntu-latest'
+    steps:
+      - uses: 'actions/checkout@master'
+      - name: 'TODO to Issue'
+        uses: 'alstr/todo-to-issue-action@master'
+        with:
+          REPO: ${{ github.repository }}
+          BEFORE: ${{ github.event.before }}
+          SHA: ${{ github.sha }}
+          TOKEN: ${{ secrets.GITHUB_TOKEN }}
+          LABEL: 'TODO'
+        id: 'todo'
