+++ .idea/runConfigurations/Deno__main_ts.xml
@@ -1,5 +1,5 @@
 <component name="ProjectRunConfigurationManager">
-  <configuration default="false" name="Deno: main.ts" type="DenoConfigurationType" inputPath="$PROJECT_DIR$/main.ts" programParameters="run --unstable --allow-net --allow-run --lock=lock.json">
+  <configuration default="false" name="Deno: main.ts" type="DenoConfigurationType" inputPath="$PROJECT_DIR$/src/main.ts" programParameters="run --unstable --allow-net --allow-run --lock=lock.json">
     <option name="applicationArguments" value="$Prompt$" />
     <method v="2" />
   </configuration>
