+++ .idea/runConfigurations/deno__write_lock.xml
@@ -1,6 +1,6 @@
 <component name="ProjectRunConfigurationManager">
   <configuration default="false" name="deno: write lock" type="ShConfigurationType">
-    <option name="SCRIPT_TEXT" value="deno run --unstable --lock=lock.json --lock-write --reload deps.ts &amp;&amp; deno fmt lock.json" />
+    <option name="SCRIPT_TEXT" value="deno run --unstable --lock=lock.json --lock-write --reload src/deps.ts &amp;&amp; deno fmt lock.json" />
     <option name="INDEPENDENT_SCRIPT_PATH" value="true" />
     <option name="SCRIPT_PATH" value="" />
     <option name="SCRIPT_OPTIONS" value="" />
